wip: Separado arquivos do contex e config
